requirements.txt
flask>=3.0.0
flask-limiter>=3.5.0
flask-login>=0.6.3
flask-wtf>=1.2.0
flask-babel>=4.0.0
google-generativeai>=0.8.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
gunicorn>=22.0.0
whitenoise>=6.7.0
pywebpush>=2.0.0
apscheduler>=3.10.0
anthropic>=0.39.0
openai>=1.50.0
pytest>=8.0.0
pytest-flask>=1.3.0
pytest-cov>=5.0.0
ruff>=0.8.0
tenacity>=8.2.0
